requirements.txt
# ─────────────────────────────────────────────
#  AI Shorts Generator – Python dependencies
#  Requires Python 3.11  (spaCy/blis incompatible with 3.13)
# ─────────────────────────────────────────────

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Rate limiting
slowapi==0.1.9

# YouTube download
yt-dlp==2026.3.3

# Audio transcription (local, no API key)
faster-whisper==1.1.1

# Video processing
moviepy==2.1.1
ffmpeg-python==0.2.0

# AI / NLP
spacy>=3.0.0
# After install run: python -m spacy download en_core_web_sm
requests==2.31.0          # Ollama HTTP calls

# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
aiofiles==23.2.1
httpx==0.27.0
tqdm==4.66.2
numpy>=2.0.0          # Updated for opencv-python 4.13+ compatibility
Pillow==10.3.0
imageio==2.34.1
imageio-ffmpeg==0.5.1
decorator==4.4.2

# Database
sqlalchemy==2.0.30
aiosqlite==0.20.0         # SQLite async driver (dev)
asyncpg==0.29.0           # PostgreSQL async driver (prod)
psycopg2-binary==2.9.9    # PostgreSQL sync driver (alembic migrations)
alembic==1.13.1           # DB migrations

# Auth & email
python-jose[cryptography]==3.3.0  # JWT encode/decode
bcrypt==4.1.3                     # password hashing
passlib[bcrypt]==1.7.4            # password hashing helper
authlib==1.3.1                    # Google OAuth
stripe==9.9.0                     # Stripe payments
fastapi-mail==1.4.1
groq>=1.1.0

# Queue & Job Processing (Advanced Twitch)
redis>=4.5.0              # Redis client
rq>=1.13.0                # RQ job queue (simple alternative to Celery)
librosa>=0.10.0           # Audio feature extraction
scipy>=1.11.0             # Required by librosa
scikit-learn>=1.3.0       # For advanced audio analysis
